requirements.txt
python-chess
cairosvg
tqdm
pydub
